refactor: move setReservationStatus Method Related: #10
