+++ src/main/java/com/modoospace/reservation/controller/dto/ReservationCreateDto.java
@@ -2,9 +2,7 @@ package com.modoospace.reservation.controller.dto;
 
 import com.modoospace.member.domain.Member;
 import com.modoospace.reservation.domain.Reservation;
-import com.modoospace.reservation.domain.ReservationStatus;
 import com.modoospace.space.domain.Facility;
-import com.modoospace.space.domain.FacilityType;
 import java.time.LocalDateTime;
 import javax.validation.constraints.NotBlank;
 import lombok.Builder;
@@ -28,26 +26,12 @@ public class ReservationCreateDto {
     this.reservationEnd = reservationEnd;
   }
 
-  public Reservation toEntity(Facility facility, Member visitor) {
-    ReservationStatus reservationStatus = setReservationStatusByFacilityType(facility.getFacilityType());
-
+  public Reservation toEntity(Facility facility, Member visitor){
     return Reservation.builder()
         .reservationStart(reservationStart)
         .reservationEnd(reservationEnd)
-        .status(reservationStatus)
         .visitor(visitor)
         .facility(facility)
         .build();
   }
-
-  private ReservationStatus setReservationStatusByFacilityType(FacilityType facilityType) {
-    if (facilityType == FacilityType.SEAT) {
-      return ReservationStatus.COMPLETED;
-    } else if (facilityType == FacilityType.ROOM) {
-      return ReservationStatus.WAITING;
-    } else {
-      return ReservationStatus.WAITING;
-    }
-  }
-
 }
