+++ src/main/java/com/modoospace/reservation/domain/ReservationRepository.java
@@ -3,7 +3,6 @@ package com.modoospace.reservation.domain;
 import com.modoospace.member.domain.Member;
 import com.modoospace.reservation.controller.dto.ReservationReadDto;
 import java.util.List;
-import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ReservationRepository extends JpaRepository<Reservation, Long> {
